[SIDE-1196] Refactor menu components
