+++ docs/src/components/menu/menu.module.scss
@@ -78,13 +78,3 @@
 .sectionSubItem {
   padding: 5px 0;
 }
-
-.mobileNavigation {
-  flex: 0 0 100px;
-}
-
-.mobileTitleMark {
-  flex: 1 1 200px;
-  font-size: $bodyFontSize;
-  margin: 0;
-}
