+++ docs/src/components/menu/menu.tsx
@@ -1,119 +1,9 @@
-import React, { ChangeEvent } from 'react';
-import {
-  buildAptitudeId,
-  buildAptitudePath,
-  buildCapabilityPath,
-} from '../aptitudes/aptitudePaths';
-import styles from './menu.module.scss';
-import { graphql, Link, navigate, StaticQuery } from 'gatsby';
-import { aptitudes, IAptitudeData } from '../aptitudes/aptitudeData';
-import { OliveHelpsLogoSmall } from '../olive-helps-logo-small';
-import { IGuideFrontMatter, IMarkdownRemarkQuery } from '../../../gatsby-node';
-
-interface IMenuProps {
-  currentPath: string;
-}
-
-type IGuideQuery = IMarkdownRemarkQuery<IGuideFrontMatter>;
-
-interface IMenuDetailProps extends IMenuProps {
-  aptitudes: IAptitudeData[];
-  guideList: IGuideQuery;
-}
-
-interface IMenuAptitudeProps {
-  aptitude: IAptitudeData;
-  current: boolean;
-}
-
-export const getPages = (queryResults: IGuideQuery): IGuideFrontMatter[] => {
-  return queryResults.allMarkdownRemark.edges.map((edge) => ({
-    slug: '/' + edge.node.frontmatter.slug,
-    title: edge.node.frontmatter.title,
-  }));
-};
-
-export const MenuAptitude: React.FunctionComponent<IMenuAptitudeProps> = (props) => {
-  const sensor = props.aptitude;
-  const capabilities = sensor.capabilities.map((capability) => {
-    return (
-      <li className={styles.sectionSubItem}>
-        <Link to={buildCapabilityPath(capability, sensor)}>{capability.name}</Link>
-      </li>
-    );
-  });
-
-  return (
-    <li className={styles.sectionItem}>
-      <Link to={buildAptitudePath(sensor)}>
-        <h2 className={styles.sectionItemHeader}>{sensor.name}</h2>
-      </Link>
-      {props.current && <ul className={styles.sectionSubItems}>{capabilities}</ul>}
-    </li>
-  );
-};
-
-export const MobileMenuSelect: React.FunctionComponent<IMenuDetailProps> = (props) => {
-  const onChange = (newValue: ChangeEvent<HTMLSelectElement>) => {
-    navigate(newValue.target.value);
-  };
-  const sensorOptions = props.aptitudes.map((apt) => (
-    <option value={buildAptitudePath(apt)}>{apt.name}</option>
-  ));
-  const guides = getPages(props.guideList).map((guide) => (
-    <option value={guide.slug}>{guide.title}</option>
-  ));
-  return (
-    <div className={styles.mobileMenu}>
-      <h1 className={styles.mobileTitleMark}>Olive Helps Developer Hub</h1>
-      <select onChange={onChange} value={props.currentPath} className={styles.mobileNavigation}>
-        <option value="/">Home</option>
-        <optgroup label="Aptitudes">{sensorOptions}</optgroup>
-        <optgroup label="Guides">{guides}</optgroup>
-      </select>
-    </div>
-  );
-};
-
-export const DesktopMenu: React.FunctionComponent<IMenuDetailProps> = (props) => {
-  const aptitudes = props.aptitudes.map((aptitude) => {
-    const aptitudeId = buildAptitudePath(aptitude);
-    return (
-      <MenuAptitude
-        aptitude={aptitude}
-        key={aptitudeId}
-        current={props.currentPath == aptitudeId}
-      />
-    );
-  });
-  const guides = getPages(props.guideList).map((guide) => (
-    <li className={styles.sectionItem}>
-      <Link to={guide.slug}>
-        <h2 className={styles.sectionItemHeader}>{guide.title}</h2>
-      </Link>
-    </li>
-  ));
-  return (
-    <div className={styles.desktopMenu}>
-      <section className={styles.menuSection}>
-        <h1 className={styles.menuTitle}>
-          <Link to="/">
-            <OliveHelpsLogoSmall className={styles.menuLogo} />
-            Developer Hub
-          </Link>
-        </h1>
-      </section>
-      <section className={styles.menuSection}>
-        <h1 className={styles.sectionTitle}>Aptitudes</h1>
-        <ul className={styles.sectionItems}>{aptitudes}</ul>
-      </section>
-      <section className={styles.menuSection}>
-        <h1 className={styles.sectionTitle}>Guides</h1>
-        <ul className={styles.sectionItems}>{guides}</ul>
-      </section>
-    </div>
-  );
-};
+import React from 'react';
+import { graphql, StaticQuery } from 'gatsby';
+import { aptitudes } from '../aptitudes/aptitudeData';
+import { IMenuProps } from './shared-menu';
+import { DesktopMenu } from './desktop-menu';
+import { MobileMenu } from './mobile-menu';
 
 export const Menu: React.FunctionComponent<IMenuProps> = (props) => {
   const aptitudeData = Object.values(aptitudes);
@@ -136,7 +26,7 @@ export const Menu: React.FunctionComponent<IMenuProps> = (props) => {
       query={guideQuery}
       render={(data) => (
         <>
-          <MobileMenuSelect
+          <MobileMenu
             aptitudes={aptitudeData}
             currentPath={props.currentPath}
             guideList={data}
