+++ docs/src/components/menu/mobile-menu.module.scss
@@ -1,0 +1,11 @@
+@import '../../pages/global';
+
+.mobileNavigation {
+  flex: 0 0 100px;
+}
+
+.mobileTitleMark {
+  flex: 1 1 200px;
+  font-size: $bodyFontSize;
+  margin: 0;
+}
